+++ Engine/actor.c
@@ -121,7 +121,7 @@ Actor *CActor_LoadTexturedModel(int id, const char *name, void *dataStart, void 
 Actor *CActor_CreateCamera(int id, const char *name, double positionX, double positionY, double positionZ,
     double rotX, double rotY, double rotZ, double angle,
     double centerX, double centerY, double centerZ, double radius,
-    double extentX, double extentY, double extentZ, int fov, enum ColliderType collider)
+    double extentX, double extentY, double extentZ, double fov, enum ColliderType collider)
 {
     Camera *camera = (Camera *)malloc(sizeof(Camera));
     camera->fov = fov;
